requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
easyocr==1.7.2
filelock==3.16.1
fsspec==2024.10.0
imageio==2.36.0
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
ninja==1.11.1.1
numpy==2.1.3
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pyclipper==1.3.0.post6
python-bidi==0.6.3
PyYAML==6.0.2
scikit-image==0.24.0
scipy==1.14.1
shapely==2.0.6
sqlparse==0.5.1
sympy==1.13.1
tifffile==2024.9.20
torch==2.5.1
torchvision==0.20.1
typing_extensions==4.12.2
tzdata==2024.2
